requirements.txt
dipy==1.7.0
einops==0.7.0
FrEIA==0.2
matplotlib==3.8.2
nibabel==5.1.0
numpy==1.24.2
Pillow==10.1.0
pycocotools==2.0.7
scikit_learn==1.3.2
scipy==1.11.4
setuptools==65.6.3
scikit-image
tifffile==2023.9.26
torch==2.1.1
torchgeometry==0.1.2
torchinfo==1.8.0
torchvision==0.16.1
tqdm==4.65.0
wandb==0.16.1
transformers
simpleitk
pie-torch~=0.0.1
